Upgraded user autologout
